Add device type to signup args
