+++ protocol/avdl/keybase1/signup.avdl
@@ -11,7 +11,7 @@ protocol signup {
   }
 
   void checkUsernameAvailable(int sessionID, string username);
-  SignupRes signup(int sessionID, string email, string inviteCode, string passphrase, string username, string deviceName, boolean storeSecret, boolean skipMail, boolean genPGPBatch);
+  SignupRes signup(int sessionID, string email, string inviteCode, string passphrase, string username, string deviceName, DeviceType deviceType, boolean storeSecret, boolean skipMail, boolean genPGPBatch);
   void inviteRequest(int sessionID, string email, string fullname, string notes);
   void checkInvitationCode(int sessionID, string invitationCode);
 }
